Remove crappy CSRF protection in favor of csurf
I blame OWASP for making it seem easy. With csurf it now stores a
secret on the session instead of the token. The actual token is salt +
hash(salt, secret), so the token is different for every request but we
can still verify it easily by extracting the salt and running the hash
over it and the secret again.

csurf also has some other nice features like automatically disabling
itself for GET requests and checking the right headers for
AJAX. Really the only thing that bugs me is the use of poor random
functions but it should be adequate for this purpose.